End of lecture 1.

+++ Lecture1/SomeJava.java
@@ -1,5 +1,5 @@
 class SomeJava {
 	public static void main(String[] args) {
-		int x = 4.2;
+		int x = (int) 4.2;
 	}
 }
